trunk/projetos/ap-comum/source/br/com/ap/comum/formatador/instancia/FormatadorDeShort.java
/*
 * FormatadorDeInteger.java
 * 
 * Data de criação: 25/06/2007
 *
 * 
 * 
 */
package br.com.ap.comum.formatador.instancia;

import br.com.ap.comum.excecao.FormatadorException;
import br.com.ap.comum.formatador.UtilFormatadorDeInteiro;

/**
 * Formatador default de short.
 * 
 * @author adrianop
 */
public class FormatadorDeShort extends FormatadorAbstrato<Short> {

	/**
	 * @param objeto Objeto formatado
	 * @return string formatada
	 * @throws FormatadorException Exceção de formatação
	 * @see br.com.ap.comum.formatador.instancia.IFormatador#formatar(java.lang.Object)
	 */
	public String formatar(Short objeto) throws FormatadorException {
		String resultado = getStringFactory().novaString();

		if (isReferencia(objeto)) {
			resultado = UtilFormatadorDeInteiro.formatar(objeto);
		}
		return resultado;
	}
}
